#3 refactor
